move the interleave tests to operand evaluator
