+++ rust/sedona-spatial-join/src/operand_evaluator.rs
@@ -17,6 +17,7 @@
 use core::fmt;
 use std::{mem::transmute, sync::Arc};
 
+use arrow::compute::interleave as arrow_interleave;
 use arrow_array::{Array, ArrayRef, Float64Array, RecordBatch};
 use arrow_schema::DataType;
 use datafusion_common::{utils::proxy::VecAllocExt, JoinSide, Result, ScalarValue};
@@ -155,6 +156,141 @@ impl EvaluatedGeometryArray {
         // for the lifetime of the GeometryBatchResult. We shorten the lifetime of the WKBs from 'static
         // to '_, so that the caller can use the WKBs without worrying about the lifetime.
         &self.wkbs
+    }
+
+    /// Build a new `EvaluatedGeometryArray` by interleaving rows from the provided
+    /// source arrays according to `indices`. Each `(batch_idx, row_idx)` pair
+    /// identifies a source array and row.
+    ///
+    /// The rectangles are gathered directly from the source arrays (no
+    /// recomputation), while the WKB references are recomputed from the
+    /// interleaved Arrow array since their lifetimes are tied to it.
+    pub fn interleave(
+        geom_arrays: &[&EvaluatedGeometryArray],
+        indices: &[(usize, usize)],
+    ) -> Result<Self> {
+        if geom_arrays.is_empty() {
+            return sedona_internal_err!("interleave requires at least one geometry array");
+        }
+
+        let sedona_type = &geom_arrays[0].sedona_type;
+
+        // Interleave the Arrow geometry arrays.
+        let value_refs: Vec<&dyn Array> = geom_arrays
+            .iter()
+            .map(|g| g.geometry_array.as_ref())
+            .collect();
+        let geometry_array = arrow_interleave(&value_refs, indices)?;
+
+        // Gather rects by index — no recomputation.
+        let rects: Vec<Option<Rect<f32>>> = indices
+            .iter()
+            .map(|&(batch_idx, row_idx)| geom_arrays[batch_idx].rects[row_idx])
+            .collect();
+
+        // Recompute WKBs from the new geometry array (lifetimes tied to it).
+        let num_rows = geometry_array.len();
+        let mut wkbs = Vec::with_capacity(num_rows);
+        geometry_array.iter_as_wkb(sedona_type, num_rows, |wkb_opt| {
+            wkbs.push(wkb_opt);
+            Ok(())
+        })?;
+        // Safety: same justification as try_new — wkbs reference buffers owned
+        // by `geometry_array` which is owned by the returned struct.
+        let wkbs = wkbs
+            .into_iter()
+            .map(|wkb| wkb.map(|wkb| unsafe { transmute(wkb) }))
+            .collect();
+
+        // Interleave distance columns.
+        let distance = Self::interleave_distance(geom_arrays, indices)?;
+
+        Ok(Self {
+            sedona_type: sedona_type.clone(),
+            geometry_array,
+            rects,
+            distance,
+            wkbs,
+        })
+    }
+
+    /// Interleave the optional distance metadata across source geometry arrays.
+    pub(crate) fn interleave_distance(
+        geom_arrays: &[&EvaluatedGeometryArray],
+        indices: &[(usize, usize)],
+    ) -> Result<Option<ColumnarValue>> {
+        let mut first_value: Option<&ColumnarValue> = None;
+        let mut needs_array = false;
+        let mut all_null = true;
+        let mut first_scalar: Option<&ScalarValue> = None;
+
+        for geom in geom_arrays {
+            match &geom.distance {
+                Some(value) => {
+                    if first_value.is_none() {
+                        first_value = Some(value);
+                    }
+                    match value {
+                        ColumnarValue::Array(array) => {
+                            needs_array = true;
+                            if all_null && array.logical_null_count() != array.len() {
+                                all_null = false;
+                            }
+                        }
+                        ColumnarValue::Scalar(scalar) => {
+                            if let Some(first) = first_scalar {
+                                if first != scalar {
+                                    needs_array = true;
+                                }
+                            } else {
+                                first_scalar = Some(scalar);
+                            }
+                            if !scalar.is_null() {
+                                all_null = false;
+                            }
+                        }
+                    }
+                }
+                None => {
+                    if first_value.is_some() && !all_null {
+                        return sedona_internal_err!(
+                            "Inconsistent distance metadata across batches"
+                        );
+                    }
+                }
+            }
+        }
+
+        if all_null {
+            return Ok(None);
+        }
+
+        let Some(distance_value) = first_value else {
+            return Ok(None);
+        };
+
+        if !needs_array {
+            if let ColumnarValue::Scalar(value) = distance_value {
+                return Ok(Some(ColumnarValue::Scalar(value.clone())));
+            }
+        }
+
+        let mut arrays: Vec<ArrayRef> = Vec::with_capacity(geom_arrays.len());
+        for geom in geom_arrays {
+            match &geom.distance {
+                Some(ColumnarValue::Array(array)) => arrays.push(array.clone()),
+                Some(ColumnarValue::Scalar(value)) => {
+                    arrays.push(value.to_array_of_size(geom.geometry_array.len())?);
+                }
+                None => {
+                    return sedona_internal_err!("Inconsistent distance metadata across batches");
+                }
+            }
+        }
+
+        let array_refs: Vec<&dyn Array> = arrays.iter().map(|a| a.as_ref()).collect();
+        let array = arrow_interleave(&array_refs, indices)?;
+        Ok(Some(ColumnarValue::Array(array)))
     }
 
     pub fn in_mem_size(&self) -> Result<usize> {
@@ -394,5 +530,206 @@ impl OperandEvaluator for KNNOperandEvaluator {
     ) -> Result<Option<f64>> {
         // NOTE: We do not support distance-based refinement for KNN predicates in the refiner phase.
         Ok(None)
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use arrow_array::BinaryArray;
+    use sedona_geometry::wkb_factory::wkb_point;
+    use sedona_schema::datatypes::WKB_GEOMETRY;
+
+    use super::*;
+
+    fn make_geom_array_with_distance(
+        wkbs: Vec<Vec<u8>>,
+        distance: Option<ColumnarValue>,
+    ) -> Result<EvaluatedGeometryArray> {
+        let geom_array: ArrayRef = Arc::new(BinaryArray::from(
+            wkbs.iter()
+                .map(|wkb| Some(wkb.as_slice()))
+                .collect::<Vec<_>>(),
+        ));
+        let mut geom = EvaluatedGeometryArray::try_new(geom_array, &WKB_GEOMETRY)?;
+        geom.distance = distance;
+        Ok(geom)
+    }
+
+    #[test]
+    fn interleave_distance_none() -> Result<()> {
+        let wkbs1 = vec![
+            wkb_point((10.0, 10.0)).unwrap(),
+            wkb_point((20.0, 20.0)).unwrap(),
+        ];
+        let wkbs2 = vec![wkb_point((30.0, 30.0)).unwrap()];
+
+        let geom1 = make_geom_array_with_distance(wkbs1, None)?;
+        let geom2 = make_geom_array_with_distance(wkbs2, None)?;
+
+        let geom_arrays = vec![&geom1, &geom2];
+        let assignments = vec![(0, 0), (1, 0), (0, 1)];
+
+        let result = EvaluatedGeometryArray::interleave_distance(&geom_arrays, &assignments)?;
+        assert!(result.is_none());
+        Ok(())
+    }
+
+    #[test]
+    fn interleave_distance_uniform_scalar() -> Result<()> {
+        let wkbs1 = vec![
+            wkb_point((10.0, 10.0)).unwrap(),
+            wkb_point((20.0, 20.0)).unwrap(),
+        ];
+        let wkbs2 = vec![wkb_point((30.0, 30.0)).unwrap()];
+
+        let scalar = ScalarValue::Float64(Some(5.0));
+        let geom1 =
+            make_geom_array_with_distance(wkbs1, Some(ColumnarValue::Scalar(scalar.clone())))?;
+        let geom2 =
+            make_geom_array_with_distance(wkbs2, Some(ColumnarValue::Scalar(scalar.clone())))?;
+
+        let geom_arrays = vec![&geom1, &geom2];
+        let assignments = vec![(0, 0), (1, 0), (0, 1)];
+
+        let result = EvaluatedGeometryArray::interleave_distance(&geom_arrays, &assignments)?;
+        assert!(matches!(result, Some(ColumnarValue::Scalar(_))));
+        if let Some(ColumnarValue::Scalar(value)) = result {
+            assert_eq!(value, ScalarValue::Float64(Some(5.0)));
+        }
+        Ok(())
+    }
+
+    #[test]
+    fn interleave_distance_different_scalars() -> Result<()> {
+        use arrow_array::Float64Array;
+
+        let wkbs1 = vec![
+            wkb_point((10.0, 10.0)).unwrap(),
+            wkb_point((20.0, 20.0)).unwrap(),
+        ];
+        let wkbs2 = vec![wkb_point((30.0, 30.0)).unwrap()];
+
+        let scalar1 = ScalarValue::Float64(Some(5.0));
+        let scalar2 = ScalarValue::Float64(Some(10.0));
+        let geom1 = make_geom_array_with_distance(wkbs1, Some(ColumnarValue::Scalar(scalar1)))?;
+        let geom2 = make_geom_array_with_distance(wkbs2, Some(ColumnarValue::Scalar(scalar2)))?;
+
+        let geom_arrays = vec![&geom1, &geom2];
+        let assignments = vec![(0, 0), (1, 0), (0, 1)];
+
+        let result = EvaluatedGeometryArray::interleave_distance(&geom_arrays, &assignments)?;
+        assert!(matches!(result, Some(ColumnarValue::Array(_))));
+        if let Some(ColumnarValue::Array(array)) = result {
+            let float_array = array.as_any().downcast_ref::<Float64Array>().unwrap();
+            assert_eq!(float_array.len(), 3);
+            assert_eq!(float_array.value(0), 5.0);
+            assert_eq!(float_array.value(1), 10.0);
+            assert_eq!(float_array.value(2), 5.0);
+        }
+        Ok(())
+    }
+
+    #[test]
+    fn interleave_distance_arrays() -> Result<()> {
+        use arrow_array::Float64Array;
+
+        let wkbs1 = vec![
+            wkb_point((10.0, 10.0)).unwrap(),
+            wkb_point((20.0, 20.0)).unwrap(),
+        ];
+        let wkbs2 = vec![wkb_point((30.0, 30.0)).unwrap()];
+
+        let array1: ArrayRef = Arc::new(Float64Array::from(vec![1.0, 2.0]));
+        let array2: ArrayRef = Arc::new(Float64Array::from(vec![3.0]));
+        let geom1 = make_geom_array_with_distance(wkbs1, Some(ColumnarValue::Array(array1)))?;
+        let geom2 = make_geom_array_with_distance(wkbs2, Some(ColumnarValue::Array(array2)))?;
+
+        let geom_arrays = vec![&geom1, &geom2];
+        let assignments = vec![(0, 0), (1, 0), (0, 1)];
+
+        let result = EvaluatedGeometryArray::interleave_distance(&geom_arrays, &assignments)?;
+        assert!(matches!(result, Some(ColumnarValue::Array(_))));
+        if let Some(ColumnarValue::Array(array)) = result {
+            let float_array = array.as_any().downcast_ref::<Float64Array>().unwrap();
+            assert_eq!(float_array.len(), 3);
+            assert_eq!(float_array.value(0), 1.0);
+            assert_eq!(float_array.value(1), 3.0);
+            assert_eq!(float_array.value(2), 2.0);
+        }
+        Ok(())
+    }
+
+    #[test]
+    fn interleave_distance_mixed_scalar_and_array() -> Result<()> {
+        use arrow_array::Float64Array;
+
+        let wkbs1 = vec![
+            wkb_point((10.0, 10.0)).unwrap(),
+            wkb_point((20.0, 20.0)).unwrap(),
+        ];
+        let wkbs2 = vec![wkb_point((30.0, 30.0)).unwrap()];
+
+        let scalar = ScalarValue::Float64(Some(5.0));
+        let array: ArrayRef = Arc::new(Float64Array::from(vec![10.0]));
+        let geom1 = make_geom_array_with_distance(wkbs1, Some(ColumnarValue::Scalar(scalar)))?;
+        let geom2 = make_geom_array_with_distance(wkbs2, Some(ColumnarValue::Array(array)))?;
+
+        let geom_arrays = vec![&geom1, &geom2];
+        let assignments = vec![(0, 0), (1, 0), (0, 1)];
+
+        let result = EvaluatedGeometryArray::interleave_distance(&geom_arrays, &assignments)?;
+        assert!(matches!(result, Some(ColumnarValue::Array(_))));
+        if let Some(ColumnarValue::Array(array)) = result {
+            let float_array = array.as_any().downcast_ref::<Float64Array>().unwrap();
+            assert_eq!(float_array.len(), 3);
+            assert_eq!(float_array.value(0), 5.0);
+            assert_eq!(float_array.value(1), 10.0);
+            assert_eq!(float_array.value(2), 5.0);
+        }
+        Ok(())
+    }
+
+    #[test]
+    fn interleave_distance_inconsistent_metadata() -> Result<()> {
+        let wkbs1 = vec![wkb_point((10.0, 10.0)).unwrap()];
+        let wkbs2 = vec![wkb_point((20.0, 20.0)).unwrap()];
+
+        let scalar = ScalarValue::Float64(Some(5.0));
+        let geom1 = make_geom_array_with_distance(wkbs1, Some(ColumnarValue::Scalar(scalar)))?;
+        let geom2 = make_geom_array_with_distance(wkbs2, None)?;
+
+        let geom_arrays = vec![&geom1, &geom2];
+        let assignments = vec![(0, 0), (1, 0)];
+
+        let result = EvaluatedGeometryArray::interleave_distance(&geom_arrays, &assignments);
+        assert!(result.is_err());
+        if let Err(e) = result {
+            assert!(e.to_string().contains("Inconsistent distance metadata"));
+        }
+        Ok(())
+    }
+
+    #[test]
+    fn interleave_distance_mixed_none_and_null() -> Result<()> {
+        use arrow_array::Float64Array;
+
+        let wkbs1 = vec![wkb_point((10.0, 10.0)).unwrap()];
+        let wkbs2 = vec![wkb_point((20.0, 20.0)).unwrap()];
+        let wkbs3 = vec![wkb_point((30.0, 30.0)).unwrap()];
+
+        let null_array = Arc::new(Float64Array::new_null(1));
+        let ega1 = make_geom_array_with_distance(wkbs1, Some(ColumnarValue::Array(null_array)))?;
+
+        let null_scalar = ScalarValue::Float64(None);
+        let ega2 = make_geom_array_with_distance(wkbs2, Some(ColumnarValue::Scalar(null_scalar)))?;
+
+        let ega3 = make_geom_array_with_distance(wkbs3, None)?;
+
+        let vec_ega = vec![&ega1, &ega2, &ega3];
+        let assignments = vec![(0, 0), (1, 0), (2, 0)];
+
+        let result = EvaluatedGeometryArray::interleave_distance(&vec_ega, &assignments)?;
+        assert!(result.is_none());
+        Ok(())
     }
 }
